+++ src/safeds/data/tabular/containers/_table.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import copy
 import functools
 import io
 import warnings
@@ -38,6 +37,9 @@ if TYPE_CHECKING:
     from safeds.data.tabular.transformation import InvertibleTableTransformer, TableTransformer
 
     from ._tagged_table import TaggedTable
+
+# Enable copy-on-write for pandas dataframes
+pd.options.mode.copy_on_write = True
 
 
 # noinspection PyProtectedMember
@@ -420,7 +422,7 @@ class Table:
         self._data = self._data.reset_index(drop=True)
         self._schema: Schema = Schema._from_pandas_dataframe(self._data)
 
-    def __eq__(self, other: Any) -> bool:
+    def __eq__(self, other: object) -> bool:
         """
         Compare two table instances.
 
@@ -469,12 +471,12 @@ class Table:
         >>> repr(table)
         '   a  b\n0  1  2\n1  3  4'
         """
-        tmp = self._data.copy(deep=True)
+        tmp = self._data.reset_index(drop=True)
         tmp.columns = self.column_names
         return tmp.__repr__()
 
     def __str__(self) -> str:
-        tmp = self._data.copy(deep=True)
+        tmp = self._data.reset_index(drop=True)
         tmp.columns = self.column_names
         return tmp.__str__()
 
@@ -831,12 +833,12 @@ class Table:
     # Transformations
     # ------------------------------------------------------------------------------------------------------------------
 
-    # This method is meant as a way to "cast" instances of subclasses of `Table` to a proper `Table`, dropping any
-    # additional constraints that might have to hold in the subclass.
-    # Override accordingly in subclasses.
     def _as_table(self: Table) -> Table:
         """
         Transform the table to an instance of the Table class.
+
+        This method is meant as a way to "cast" instances of subclasses of `Table` to a proper `Table`, dropping any
+        additional constraints that might have to hold in the subclass. Override accordingly in subclasses.
 
         Returns
         -------
@@ -879,7 +881,7 @@ class Table:
         if column.number_of_rows != self.number_of_rows and self.number_of_columns != 0:
             raise ColumnSizeError(str(self.number_of_rows), str(column._data.size))
 
-        result = self._data.copy()
+        result = self._data.reset_index(drop=True)
         result.columns = self._schema.column_names
         result[column.name] = column._data
         return Table._from_pandas_dataframe(result)
@@ -920,7 +922,7 @@ class Table:
         """
         if isinstance(columns, Table):
             columns = columns.to_columns()
-        result = self._data.copy()
+        result = self._data.reset_index(drop=True)
         result.columns = self._schema.column_names
         for column in columns:
             if column.name in result.columns:
@@ -969,7 +971,7 @@ class Table:
         1  3  4
         """
         int_columns = []
-        result = self._copy()
+
         if self.number_of_columns == 0:
             return Table.from_rows([row])
         if len(set(self.column_names) - set(row.column_names)) > 0:
@@ -980,12 +982,12 @@ class Table:
                 ),
             )
 
-        if result.number_of_rows == 0:
+        if self.number_of_rows == 0:
             int_columns = list(filter(lambda name: isinstance(row[name], int | np.int64 | np.int32), row.column_names))
 
-        new_df = pd.concat([result._data, row._data]).infer_objects()
-        new_df.columns = result.column_names
-        schema = Schema.merge_multiple_schemas([result.schema, row.schema])
+        new_df = pd.concat([self._data, row._data]).infer_objects()
+        new_df.columns = self.column_names
+        schema = Schema.merge_multiple_schemas([self.schema, row.schema])
         result = Table._from_pandas_dataframe(new_df, schema)
 
         for column in int_columns:
@@ -1033,7 +1035,7 @@ class Table:
             rows = rows.to_rows()
 
         if len(rows) == 0:
-            return self._copy()
+            return self
 
         different_column_names = set()
         for row in rows:
@@ -1046,8 +1048,7 @@ class Table:
                 ),
             )
 
-        result = self._copy()
-
+        result = self
         for row in rows:
             result = result.add_row(row)
 
@@ -1153,9 +1154,7 @@ class Table:
         if len(invalid_columns) != 0:
             raise UnknownColumnNameError(invalid_columns, similar_columns)
 
-        clone = self._copy()
-        clone = clone.remove_columns(list(set(self.column_names) - set(column_names)))
-        return clone
+        return self.remove_columns(list(set(self.column_names) - set(column_names)))
 
     def remove_columns(self, column_names: list[str]) -> Table:
         """
@@ -1309,8 +1308,7 @@ class Table:
              a    b
         0  1.0  2.0
         """
-        result = self._data.copy(deep=True)
-        result = result.dropna(axis="index")
+        result = self._data.dropna(axis="index")
         return Table._from_pandas_dataframe(result)
 
     def remove_rows_with_outliers(self) -> Table:
@@ -1350,13 +1348,11 @@ class Table:
         9   0.0  0.00  0.0 -1000000
         10  0.0  0.00  0.0 -1000000
         """
-        copy = self._data.copy(deep=True)
-
         table_without_nonnumericals = self.remove_columns_with_non_numerical_values()
         z_scores = np.absolute(stats.zscore(table_without_nonnumericals._data, nan_policy="omit"))
         filter_ = ((z_scores < 3) | np.isnan(z_scores)).all(axis=1)
 
-        return Table._from_pandas_dataframe(copy[filter_], self._schema)
+        return Table._from_pandas_dataframe(self._data[filter_], self._schema)
 
     def rename_column(self, old_name: str, new_name: str) -> Table:
         """
@@ -1399,7 +1395,7 @@ class Table:
         if new_name in self._schema.column_names:
             raise DuplicateColumnNameError(new_name)
 
-        new_df = self._data.copy()
+        new_df = self._data.reset_index(drop=True)
         new_df.columns = self._schema.column_names
         return Table._from_pandas_dataframe(new_df.rename(columns={old_name: new_name}))
 
@@ -2079,7 +2075,7 @@ class Table:
         """
         col_wrap = min(self.number_of_columns, 3)
 
-        data = pd.melt(self._data.applymap(lambda value: str(value)), value_vars=self.column_names)
+        data = pd.melt(self._data.map(lambda value: str(value)), value_vars=self.column_names)
         grid = sns.FacetGrid(data=data, col="variable", col_wrap=col_wrap, sharex=False, sharey=False)
         grid.map(sns.histplot, "value")
         grid.set_xlabels("")
@@ -2128,7 +2124,7 @@ class Table:
         if path.suffix != ".csv":
             raise WrongFileExtensionError(path, ".csv")
         path.parent.mkdir(parents=True, exist_ok=True)
-        data_to_csv = self._data.copy()
+        data_to_csv = self._data.reset_index(drop=True)
         data_to_csv.columns = self._schema.column_names
         data_to_csv.to_csv(path, index=False)
 
@@ -2166,7 +2162,7 @@ class Table:
         tmp_table_file.save(path)
 
         path.parent.mkdir(parents=True, exist_ok=True)
-        data_to_excel = self._data.copy()
+        data_to_excel = self._data.reset_index(drop=True)
         data_to_excel.columns = self._schema.column_names
         data_to_excel.to_excel(path)
 
@@ -2197,7 +2193,7 @@ class Table:
         if path.suffix != ".json":
             raise WrongFileExtensionError(path, ".json")
         path.parent.mkdir(parents=True, exist_ok=True)
-        data_to_json = self._data.copy()
+        data_to_json = self._data.reset_index(drop=True)
         data_to_json.columns = self._schema.column_names
         data_to_json.to_json(path)
 
@@ -2333,21 +2329,6 @@ class Table:
         if not allow_copy:
             raise NotImplementedError("For the moment we need to copy the data, so `allow_copy` must be True.")
 
-        data_copy = self._data.copy()
+        data_copy = self._data.reset_index(drop=True)
         data_copy.columns = self.column_names
         return data_copy.__dataframe__(nan_as_null, allow_copy)
-
-    # ------------------------------------------------------------------------------------------------------------------
-    # Helpers
-    # ------------------------------------------------------------------------------------------------------------------
-
-    def _copy(self) -> Table:
-        """
-        Return a copy of this table.
-
-        Returns
-        -------
-        table : Table
-            The copy of this table.
-        """
-        return copy.deepcopy(self)