+++ src/safeds/data/tabular/transformation/_range_scaler.py
@@ -151,7 +151,7 @@ class RangeScaler(InvertibleTableTransformer):
                 ),
             )
 
-        data = table._data.copy()
+        data = table._data.reset_index(drop=True)
         data.columns = table.column_names
         data[self._column_names] = self._wrapped_transformer.transform(data[self._column_names])
         return Table._from_pandas_dataframe(data)
@@ -213,7 +213,7 @@ class RangeScaler(InvertibleTableTransformer):
                 ),
             )
 
-        data = transformed_table._data.copy()
+        data = transformed_table._data.reset_index(drop=True)
         data.columns = transformed_table.column_names
         data[self._column_names] = self._wrapped_transformer.inverse_transform(data[self._column_names])
         return Table._from_pandas_dataframe(data)