+++ tests/safeds/data/tabular/containers/_table/test_plot_correlation_heatmap.py
@@ -1,29 +1,23 @@
 import pytest
-from safeds.data.image.containers import ImagePil
 from safeds.data.tabular.containers import Table
-
-from tests.helpers import resolve_resource_path
+from syrupy import SnapshotAssertion
 
 
 @pytest.mark.parametrize(
-    ("table", "path"),
+    "table",
     [
-        (Table({"A": [1, 2, 3.5], "B": [0.2, 4, 77]}), "./image/snapshot_heatmap.png"),
-        (Table(), "./image/snapshot_empty_heatmap.png"),
+        Table({"A": [1, 2, 3.5], "B": [0.2, 4, 77]}),
     ],
-    ids=["normal", "empty"],
+    ids=["normal"],
 )
-def test_should_match_snapshot(table: Table, path: str) -> None:
-    if table.number_of_rows == 0:
-        with pytest.warns(
-            UserWarning,
-            match=r"An empty table has been used. A correlation heatmap on an empty table will show nothing.",
-        ):
-            current = table.plot_correlation_heatmap()
-    else:
-        current = table.plot_correlation_heatmap()
-    snapshot = ImagePil.from_png_file(resolve_resource_path(path))
+def test_should_match_snapshot(table: Table, snapshot_png: SnapshotAssertion) -> None:
+    correlation_heatmap = table.plot_correlation_heatmap()
+    assert correlation_heatmap == snapshot_png
 
-    # Inlining the expression into the assert causes pytest to hang if the assertion fails when run from PyCharm.
-    assertion = snapshot == current
-    assert assertion
+
+def test_should_warn_about_empty_table() -> None:
+    with pytest.warns(
+        UserWarning,
+        match=r"An empty table has been used. A correlation heatmap on an empty table will show nothing.",
+    ):
+        Table().plot_correlation_heatmap()