+++ tests/safeds/data/tabular/containers/_table/test_plot_scatterplot.py
@@ -1,19 +1,13 @@
 import pytest
-from safeds.data.image.containers import ImagePil
 from safeds.data.tabular.containers import Table
 from safeds.exceptions import UnknownColumnNameError
+from syrupy import SnapshotAssertion
 
-from tests.helpers import resolve_resource_path
 
-
-def test_should_match_snapshot() -> None:
+def test_should_match_snapshot(snapshot_png: SnapshotAssertion) -> None:
     table = Table({"A": [1, 2, 3], "B": [2, 4, 7]})
-    current = table.plot_scatterplot("A", "B")
-    snapshot = ImagePil.from_png_file(resolve_resource_path("./image/snapshot_scatterplot.png"))
-
-    # Inlining the expression into the assert causes pytest to hang if the assertion fails when run from PyCharm.
-    assertion = snapshot._image.tobytes() == current._image.tobytes()
-    assert assertion
+    scatterplot = table.plot_scatterplot("A", "B")
+    assert scatterplot == snapshot_png
 
 
 @pytest.mark.parametrize(